Split in-core into refine and is-authorized
