+++ packages/jam/in-core/refine.test.ts
@@ -1,0 +1,7 @@
+import { describe } from "node:test";
+
+describe("Refine", () => {
+  // TODO [ToDr] Add refine-specific PVM invocation tests.
+  // These should test Refine.invoke() directly, similar to
+  // how is-authorized.test.ts tests IsAuthorized.invoke().
+});
